+++ src/replace.js
@@ -1,15 +1,9 @@
-export function replace(
+import { curry } from './curry'
+
+function replaceFn(
   pattern, replacer, str
 ){
-  if (replacer === undefined){
-    return (_replacer, _str) => replace(
-      pattern, _replacer, _str
-    )
-  } else if (str === undefined){
-    return _str => replace(
-      pattern, replacer, _str
-    )
-  }
-
   return str.replace(pattern, replacer)
 }
+
+export const replace = curry(replaceFn)